+++ lib/couch_potato/view/base_view_spec.rb
@@ -5,6 +5,7 @@ module CouchPotato
       private :klass, :options
 
       def initialize(klass, view_name, options, view_parameters)
+        assert_valid_view_parameters view_parameters
         @klass = klass
         @design_document = klass.to_s.underscore
         @view_name = view_name
@@ -18,6 +19,18 @@ module CouchPotato
 
       def process_results(results)
         results
+      end
+      
+      private
+      
+      def assert_valid_view_parameters(params)
+        params.keys.each do |key|
+          raise ArgumentError.new("invalid view parameter: #{key}") unless valid_view_parameters.include?(key.to_s)
+        end
+      end
+      
+      def valid_view_parameters
+        %w(key startkey startkey_docid endkey endkey_docid limit stale descending skip group group_level reduce include_docs inclusive_end)
       end
     end
   end