+++ CHANGES.md
@@ -3,6 +3,7 @@
 ### master
 * automatically set a database instance on results of CouchPotato::Database#view (langalex)
 * improved auto loading of unloaded constants - can now load constants that have never been loaded before (langalex)
+* raise exception on invalid parameters passed to a couchdb view query (langalex)
 
 ### 0.2.20
 * support for :boolean properties (jweiss)